Rename etf-analysis-data to etf-analysis and move category-flags into it
