+++ insights-ui/src/etf-analysis/category-flags/allocation-target-date.json
