+++ insights-ui/src/etf-analysis/category-flags/broad-equity.json
