+++ insights-ui/src/etf-analysis/category-flags/commodities-and-digital-assets.json
