+++ insights-ui/src/etf-analysis/category-flags/derivative-income.json
