+++ insights-ui/src/etf-analysis/category-flags/fixed-income-credit-and-income.json
