+++ insights-ui/src/etf-analysis/category-flags/fixed-income-investment-grade.json
