+++ insights-ui/src/etf-analysis/category-flags/leveraged-inverse.json
