+++ insights-ui/src/etf-analysis/category-flags/sector-thematic-equity.json
