+++ insights-ui/src/etf-analysis/etf-analysis-categories.json
