+++ insights-ui/src/etf-analysis/etf-category-aliases.json
