+++ insights-ui/src/etf-analysis/sample-etfs.json
